+++ examples/feature1/screens/feature1_screen.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/cupertino.dart';
+
+class Feature1Screen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: Text(
+          'Feature1Screen is working', 
+        ),
+      ),
+    );
+  }
+}