+++ examples/feature1/widgets/feature1_widget.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/cupertino.dart';
+
+class Feature1Widget extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Center(
+        child: Text(
+          'Feature1Widget is working', 
+        ),
+      ),
+    );
+  }
+}