+++ lib/feature_folder_cli.dart
@@ -3,9 +3,12 @@ import 'dart:io';
 import 'package:args/command_runner.dart';
 
 import 'commands/generate.dart';
+import 'utils/logs.dart';
 
 class FeatureFolderCLI {
   static void main(List<String> args) {
+    final time = Stopwatch();
+    time.start();
     final _runner = CommandRunner('ff', 'Generate folder by feature structure.')
       ..addCommand(GenerateCommand());
     _runner.run(args).catchError((err) {
@@ -17,5 +20,7 @@ class FeatureFolderCLI {
       }
       exit(64);
     });
+    time.stop();
+    LogService.info('Time Taken: ${time.elapsed.inMilliseconds} milliseconds');
   }
 }